Works for Android
Updated the tutorial program so that you can move the character using
the accelerometers and tap the screen to fire. As of right now there's
only one monster solo because I did all of this stuff for testing and
haven't actually worked on expanding anything into a full length like
game.